+++ src/main/java/com/appdynamics/extensions/ABaseMonitor.java
@@ -15,7 +15,11 @@
 
 package com.appdynamics.extensions;
 
-import com.appdynamics.extensions.conf.MonitorConfiguration;
+import com.appdynamics.extensions.conf.MonitorContextConfiguration;
+import com.appdynamics.extensions.conf.MonitorContext;
+import com.appdynamics.extensions.file.FileWatchListener;
+import com.appdynamics.extensions.util.AssertUtils;
+import com.appdynamics.extensions.util.PathResolver;
 import com.singularity.ee.agent.systemagent.api.AManagedMonitor;
 import com.singularity.ee.agent.systemagent.api.TaskExecutionContext;
 import com.singularity.ee.agent.systemagent.api.TaskOutput;
@@ -35,7 +39,7 @@ import java.util.Map;
  * <p>An {@code ABaseMonitor} is a wrapper on top of
  * {@link AManagedMonitor} to remove the boiler plate code of
  * creating a {@link AMonitorJob} and initializing the
- * {@link MonitorConfiguration}.
+ * {@link MonitorContextConfiguration}.
  *
  * <p>The MA or SIM agent loads all the {@link AManagedMonitor}s
  * from their respective subdirectories in the monitors directory
@@ -87,11 +91,13 @@ public abstract class ABaseMonitor extends AManagedMonitor{
      */
     protected String monitorName;
 
+    private File installDir;
+
     /**
-     * A configuration object that reads the monitor's config file
+     * A contextConfiguration object that reads the monitor's config file
      * and initializes the different bits required by the monitor.
      */
-    protected MonitorConfiguration configuration;
+    private MonitorContextConfiguration contextConfiguration;
 
     /**
      * A runnable which does all the leg work for fetching the
@@ -104,19 +110,37 @@ public abstract class ABaseMonitor extends AManagedMonitor{
         logger.info("Using {} Version [{}]",monitorName, getImplementationVersion());
     }
 
-    protected void initialize(Map<String, String> args) {
-        if(configuration == null){
+    protected void initialize(final Map<String, String> args) {
+        if(contextConfiguration == null){
+            installDir = getInstallDirectory();
             monitorJob = createMonitorJob();
-            MonitorConfiguration conf = new MonitorConfiguration(monitorName,getDefaultMetricPrefix(), monitorJob);
-            conf.setConfigYml(args.get("config-file"), new MonitorConfiguration.FileWatchListener() {
-                @Override
-                public void onFileChange(File file) {
-                    onConfigReload(file);
-                }
-            });
-            initializeMoreStuff(args, conf);
-            this.configuration = conf;
+            contextConfiguration = createContextConfiguration();
+            contextConfiguration.registerListener(args.get("config-file"), createYmlFileListener(args.get("config-file")));
+            initializeMoreStuff(args);
         }
+    }
+
+    private FileWatchListener createYmlFileListener(final String ymlFile) {
+        FileWatchListener fileWatchListener = new FileWatchListener() {
+            @Override
+            public void onFileChange(File file) {
+                contextConfiguration.setConfigYml(ymlFile);
+                onConfigReload(file);
+            }
+        };
+        return fileWatchListener;
+    }
+
+    private File getInstallDirectory() {
+        File installDir = PathResolver.resolveDirectory(AManagedMonitor.class);
+        if(installDir == null){
+            throw new RuntimeException("The install directory cannot be null");
+        }
+        return installDir;
+    }
+
+    private MonitorContextConfiguration createContextConfiguration() {
+        return new MonitorContextConfiguration(getMonitorName(), getDefaultMetricPrefix(), installDir, monitorJob);
     }
 
     protected void onConfigReload(File file){};
@@ -128,9 +152,9 @@ public abstract class ABaseMonitor extends AManagedMonitor{
     /**
      * A placeholder method which should be overridden if there are
      * custom objects to be initialized in the monitor.
-     * @param conf
+     * @param args
      */
-    protected void initializeMoreStuff(Map<String, String> args, MonitorConfiguration conf) {
+    protected void initializeMoreStuff(Map<String, String> args) {
         ;
     }
 
@@ -154,8 +178,10 @@ public abstract class ABaseMonitor extends AManagedMonitor{
     }
 
     protected void executeMonitor() {
-        if(configuration.isEnabled()){
-            if(configuration.isScheduledModeEnabled()){ //scheduled mode
+        if(contextConfiguration.isEnabled()){
+            MonitorContext context = contextConfiguration.getContext();
+            AssertUtils.assertNotNull(context, "The context of the extension has not been initialised!!!! Please check your contextConfiguration");
+            if(context.isScheduledModeEnabled()){ //scheduled mode
                 logger.debug("Task scheduler is enabled, printing the metrics from the cache");
                 monitorJob.printAllFromCache();
             }
@@ -175,8 +201,8 @@ public abstract class ABaseMonitor extends AManagedMonitor{
 
     public abstract String getMonitorName();
 
-    public MonitorConfiguration getConfiguration(){
-        return configuration;
+    public MonitorContextConfiguration getContextConfiguration() {
+        return contextConfiguration;
     }
 
     protected abstract void doRun(TasksExecutionServiceProvider tasksExecutionServiceProvider);
@@ -186,4 +212,5 @@ public abstract class ABaseMonitor extends AManagedMonitor{
     protected void onComplete() {
         logger.info("Finished processing all tasks in the job for {}",getMonitorName());
     }
+
 }