+++ src/main/java/com/appdynamics/extensions/TasksExecutionServiceProvider.java
@@ -15,14 +15,9 @@
 
 package com.appdynamics.extensions;
 
-import com.appdynamics.extensions.conf.MonitorConfiguration;
-import org.apache.commons.httpclient.util.ExceptionUtil;
-import org.apache.http.util.ExceptionUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.io.PrintWriter;
-import java.io.StringWriter;
 import java.util.concurrent.atomic.AtomicInteger;
 
 /*
@@ -46,7 +41,7 @@ public class TasksExecutionServiceProvider {
 
 
     public void submit(final String name, final AMonitorTaskRunnable aServerTask){
-        aBaseMonitor.getConfiguration().getExecutorService().submit(name,new Runnable() {
+        aBaseMonitor.getContextConfiguration().getContext().getExecutorService().submit(name,new Runnable() {
             @Override
             public void run() {
                 try{
@@ -72,9 +67,5 @@ public class TasksExecutionServiceProvider {
 
     public MetricWriteHelper getMetricWriteHelper() {
         return this.metricWriteHelper;
-    }
-
-    public MonitorConfiguration getMonitorConfiguration(){
-        return this.aBaseMonitor.configuration;
     }
 }